requirements.txt
qiskit
qiskit-aer
numpy
matplotlib
pycryptodome
streamlit
